move parsing host code to new module
